requirements.txt
aiohttp~=3.7.4
discord.py~=1.7.3
asyncpg~=0.23.0
google-api-python-client~=2.14.0
dateparser~=1.0.0
praw~=7.3.0